+++ sermant-integration-tests/tag-transmission-test/jakarta-servlet6-demo/src/main/java/io/sermant/demo/tagtransmission/jakarta/servlet6/utils/TagConversionUtils.java
@@ -1,0 +1,68 @@
+/*
+ * Copyright (C) 2025-2025 Sermant Authors. All rights reserved.
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+
+package io.sermant.demo.tagtransmission.jakarta.servlet6.utils;
+
+import jakarta.servlet.http.HttpServletRequest;
+
+/**
+ * Tag transmission convert util
+ *
+ * @author daizhenyu
+ * @since 2025-05-22
+ **/
+public class TagConversionUtils {
+    private static final int BOUND_LENGTH = 2;
+
+    private static final String DOUBLE_QUOTATION_MARKS = "\"";
+
+    private TagConversionUtils() {
+    }
+
+    /**
+     * Convert the tag in the HTTP header to a JSON string
+     *
+     * @param request
+     * @param tagKeys
+     * @return string
+     */
+    public static String convertHeader2String(HttpServletRequest request, String[] tagKeys) {
+        return buildString(request, tagKeys);
+    }
+
+    private static String buildString(HttpServletRequest request, String[] tagKeys) {
+        StringBuilder builder = new StringBuilder();
+        builder.append("{");
+        for (String key : tagKeys) {
+            if (request.getHeader(key) == null) {
+                continue;
+            }
+            builder.append(DOUBLE_QUOTATION_MARKS);
+            builder.append(key);
+            builder.append(DOUBLE_QUOTATION_MARKS);
+            builder.append(":");
+            builder.append(DOUBLE_QUOTATION_MARKS);
+            builder.append(request.getHeader(key));
+            builder.append(DOUBLE_QUOTATION_MARKS);
+            builder.append(",");
+        }
+        if (builder.length() >= BOUND_LENGTH) {
+            builder.deleteCharAt(builder.length() - 1);
+        }
+        builder.append("}");
+        return builder.toString();
+    }
+}
