+++ sermant-integration-tests/tag-transmission-test/jakarta-servlet6-demo/src/main/resources/application.properties
