+++ sermant-plugins/sermant-tag-transmission/pom.xml
@@ -54,7 +54,7 @@
                 <module>tag-transmission-dubbo3.x-plugin</module>
                 <module>tag-transmission-service</module>
                 <module>tag-transmission-feign11.x-plugin</module>
-                <module>tag-transmission-servlet56-plugin</module>
+                <module>tag-transmission-jakarta-servlet-plugin</module>
             </modules>
         </profile>
         <profile>
@@ -78,7 +78,7 @@
                 <module>tag-transmission-dubbo3.x-plugin</module>
                 <module>tag-transmission-service</module>
                 <module>tag-transmission-feign11.x-plugin</module>
-                <module>tag-transmission-servlet56-plugin</module>
+                <module>tag-transmission-jakarta-servlet-plugin</module>
             </modules>
         </profile>
         <profile>
@@ -102,7 +102,7 @@
                 <module>tag-transmission-dubbo3.x-plugin</module>
                 <module>tag-transmission-service</module>
                 <module>tag-transmission-feign11.x-plugin</module>
-                <module>tag-transmission-servlet56-plugin</module>
+                <module>tag-transmission-jakarta-servlet-plugin</module>
             </modules>
         </profile>
     </profiles>
