+++ sermant-plugins/sermant-tag-transmission/tag-transmission-jakarta-servlet-plugin/src/main/resources/META-INF/services/io.sermant.core.plugin.agent.declarer.PluginDeclarer
@@ -1,0 +1,1 @@
+io.sermant.tag.transmission.jakarta.servlet.declarers.JakartaServletDeclarer
