--- sermant-plugins/sermant-tag-transmission/tag-transmission-servlet56-plugin/src/main/resources/META-INF/services/io.sermant.core.plugin.agent.declarer.PluginDeclarer
@@ -1,1 +1,0 @@
-io.sermant.tag.transmission.servlet56.declarers.Servlet56Declarer
